+++ lib/src/tile_map_parser.dart
@@ -13,6 +13,8 @@ class TileMapParser {
     var map = new TileMap();
     map.tileWidth = int.parse(xmlElement.getAttribute('tilewidth'));
     map.tileHeight = int.parse(xmlElement.getAttribute('tileheight'));
+    map.width = int.parse(xmlElement.getAttribute('width'));
+    map.height = int.parse(xmlElement.getAttribute('height'));
 
     xmlElement.children
         .where((node) => node is XmlElement)