+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 * Added typed properties support and properties for TileMap and Layer
 
+## 0.3.0
+
+* `Layer` using 2d list store layer's `_tiles`,useful for get tile by coordinate `x,y`
+* `Layer` using `px,px,x,y` instead of just `x,y`(pixel)
+* add `width`,`height` to TileMap
+
 ## 0.11.7
 
 * `<object />` tags are now parsed for their `type` attribute